project-root/storage/requirements.txt
flask==3.0.0
pyyaml==6.0.1
pillow==10.1.0